Change IRenderCore2D interface to RenderCore2D abstact base class
