+++ Source/HelixToolkit.SharpDX.Shared/HelixToolkit.SharpDX.Shared.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>HelixToolkit.SharpDX.Shared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Core2D\Abstract\RenderCore2D.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core2D\Abstract\RenderCore2DBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core2D\Abstract\ShapeRenderCore2DBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core2D\BorderRenderCore2D.cs" />
