+++ Source/HelixToolkit.SharpDX.Shared/Interface/Constants.cs
@@ -11,6 +11,7 @@ namespace HelixToolkit.Wpf.SharpDX
 #endif
 {
     using Core;
+    using Core2D;
     using System.Collections.Generic;
 
     /// <summary>
@@ -84,6 +85,6 @@ namespace HelixToolkit.Wpf.SharpDX
         public static readonly List<IRenderable> EmptyRenderable = new List<IRenderable>();
         public static readonly List<RenderCore> EmptyCore = new List<RenderCore>();
         public static readonly IList<IRenderable2D> EmptyRenderable2D = new List<IRenderable2D>();
-        public static readonly IList<IRenderCore2D> EmptyCore2D = new List<IRenderCore2D>();
+        public static readonly IList<RenderCore2D> EmptyCore2D = new List<RenderCore2D>();
     }
 }
