+++ Source/HelixToolkit.SharpDX.Shared/Interface/IRenderable2D.cs
@@ -3,8 +3,6 @@ The MIT License (MIT)
 Copyright (c) 2018 Helix Toolkit contributors
 */
 using SharpDX;
-using SharpDX.Direct2D1;
-using System;
 using System.Collections.Generic;
 
 #if NETFX_CORE
@@ -13,73 +11,7 @@ namespace HelixToolkit.UWP
 namespace HelixToolkit.Wpf.SharpDX
 #endif
 {
-    /// <summary>
-    /// 
-    /// </summary>
-    public interface IRenderCore2D
-    {        
-        /// <summary>
-        /// Occurs when [on invalidate renderer].
-        /// </summary>
-        event EventHandler<EventArgs> OnInvalidateRenderer;
-        /// <summary>
-        /// Gets or sets the bound.
-        /// </summary>
-        /// <value>
-        /// The bound.
-        /// </value>
-        RectangleF LayoutBound { set; get; }
-
-        /// <summary>
-        /// Gets or sets the clipping bound.
-        /// </summary>
-        /// <value>
-        /// The clipping bound.
-        /// </value>
-        RectangleF LayoutClippingBound { set; get; }
-        /// <summary>
-        /// 
-        /// </summary>
-        bool IsEmpty { get; }
-
-        /// <summary>
-        /// Transform to render position relative to its parent
-        /// </summary>
-        Matrix3x2 Transform { set; get; }        
-        /// <summary>
-        /// Local render transform
-        /// </summary>
-        Matrix3x2 LocalTransform { set; get; }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether this instance is rendering.
-        /// </summary>
-        /// <value>
-        ///   <c>true</c> if this instance is rendering; otherwise, <c>false</c>.
-        /// </value>
-        bool IsRendering { set; get; }
-        /// <summary>
-        /// Attaches the specified target.
-        /// </summary>
-        /// <param name="target">The target.</param>
-        void Attach(IRenderHost target);
-        /// <summary>
-        /// Detaches this instance.
-        /// </summary>
-        void Detach();
-        /// <summary>
-        /// Renders the specified context.
-        /// </summary>
-        /// <param name="context">The context.</param>
-        void Render(IRenderContext2D context);
-        /// <summary>
-        /// Gets or sets a value indicating whether this instance is mouse over.
-        /// </summary>
-        /// <value>
-        ///   <c>true</c> if this instance is mouse over; otherwise, <c>false</c>.
-        /// </value>
-        bool IsMouseOver { set; get; }
-    }
+    using Core2D;
 
     public interface IRenderable2D : ITransform2D, IGUID
     {
@@ -125,7 +57,7 @@ namespace HelixToolkit.Wpf.SharpDX
         /// </summary>
         IList<IRenderable2D> Items { get; }
 
-        IRenderCore2D RenderCore { get; }
+        RenderCore2D RenderCore { get; }
         /// <summary>
         /// Update render related parameters such as model matrix by scene graph and bounding boxes
         /// </summary>
